restructured controllers so that the admin controller handles all redirecting of admin functions, for now
