+++ app/Http/Service/AdminService.php
@@ -50,5 +50,16 @@ class AdminService
         return $lockers;
     }
 
+    public static function update_locker_status($locker_id, $status){
+        $locker = Locker::find($locker_id);
 
+        if($locker){
+            $locker->status = $status;
+            $locker->save();
+            return true;
+        }
+        else{
+            return false;
+        }
+    }
 }
